requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
flask-cors==4.0.1
werkzeug==3.1.3
sqlalchemy==2.0.44
jinja2==3.1.6
itsdangerous==2.2.0
markupsafe==3.0.3
blinker==1.9.0
typing-extensions==4.15.0
gunicorn==21.2.0
